requirements.txt
# Core dependencies
faiss-cpu>=1.7.4
numpy>=1.24
pandas>=2.0
h5py>=3.9

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Machine learning (for Phase 3 cost model)
scikit-learn>=1.3

# Configuration
pyyaml>=6.0

# System monitoring
psutil>=5.9

# Progress bars
tqdm>=4.65

# Jupyter support (optional, for notebooks)
jupyter>=1.0
ipykernel>=6.25
